backend/requirements.txt
flask
flask-cors
pillow
python-pptx
pytesseract
openai
python-dotenv
pywin32